Risk Assessment module/requirements.txt
fastapi
transformers
uvicorn
torch
openai
python-dotenv